@@ -1,24 +1,25 @@
 {
   "name": "minesweeper",
-  "version": "1.3.0",
+  "version": "1.3.1",
   "description": "good ol' windows game in react",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "del-cli dist && webpack --mode production",
+    "dev:build": "del-cli dist && webpack --mode development",
     "dev:start": "webpack serve",
     "start": "node server.js"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@types/node": "^14.17.10",
-    "@types/react": "^17.0.19",
-    "@types/react-dom": "^17.0.9",
-    "copy-webpack-plugin": "^9.0.1",
-    "del-cli": "^4.0.1",
+    "@types/node": "^16.11.56",
+    "@types/react": "^18.0.18",
+    "@types/react-dom": "^18.0.6",
+    "copy-webpack-plugin": "^11.0.0",
+    "del-cli": "^5.0.0",
     "html-webpack-plugin": "^5.3.2",
-    "source-map-loader": "^3.0.0",
+    "source-map-loader": "^4.0.0",
     "ts-loader": "^9.2.5",
     "typescript": "^4.3.5",
     "webpack": "^5.51.1",
@@ -30,8 +31,8 @@
   "dependencies": {
     "express": "^4.17.1",
     "query-string": "^7.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "keywords": []
 }
